Use unoffical ceph and nfs-ganesha packages for now
The official ceph and nfs-ganesha packages being built are too old to be
of value here, as they both lack some critical patches that we need. In
the near future that will change, but for now point to my personal COPR
repo. Eventually we should be able to revert this once the official
builds catch up.

With this change we need to alter the generated ganesha.conf files as
well:

Change to rados_cluster recovery backend as rados_kv is not suitable
for active/active deployment. Drop obsolete CACHEINODE settings, and the
RADOS_URLS config block as we aren't storing configs in RADOS.

Finally, disable delegations until we have a solution for:

    http://tracker.ceph.com/issues/24802

Signed-off-by: Jeff Layton <jlayton@redhat.com>