rbdmirror: allow deleting the peer key if not used
If the rbd mirror cluster peer keyring is not in use, allow
the user to delete the key. The keyring is only needed
in case rbd mirroring is in use, which very few clusters enable.
Since Rook does not have the context to know if other clusters
are consuming this key, Rook cannot delete it automatically.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>